+++ configs/yolov5/voc/yolov5_s-v61_fast_1xb64-50e_voc.py
@@ -206,6 +206,7 @@ custom_hooks = [
         ema_type='ExpMomentumEMA',
         momentum=0.0001,
         update_buffers=True,
+        # To load coco pretrained model, need to set `strict_load=False`
         strict_load=False,
         priority=49)
 ]
