+++ mmyolo/datasets/__init__.py
@@ -1,7 +1,8 @@
 # Copyright (c) OpenMMLab. All rights reserved.
 from .transforms import *  # noqa: F401,F403
 from .utils import BatchShapePolicy, yolov5_collate
-from .yolov5_datasets import YOLOv5CocoDataset, YOLOv5VOCDataset
+from .yolov5_coco import YOLOv5CocoDataset
+from .yolov5_voc import YOLOv5VOCDataset
 
 __all__ = [
     'YOLOv5CocoDataset', 'YOLOv5VOCDataset', 'BatchShapePolicy',
